alley: Drop allexe name in cache to focus more on content-based.
Otherwise we end up having separate cache entries for the same
code when it's in different allexes or if it's the same allexe
but with a different filename.

Still keep the entry name though since that should be the same
and is the only remaining human-readable aspect of the cache now :).