alley: Always load modules, but lazily, fix static constructors.
The way we were adding cache objects to the EE skips any ctor's
they might have had, so instead always add as a module
and then provide file to MCJIT upon request.

The reason we skipped it before was to load things faster,
which we get back much of by loading modules lazily when possible.

For some reason MCJIT gets grumpy if you don't materialize a module
it won't find in the cache, so workaround that for now.