To avoid problems getting llvm lit, just distribute our own copy.
It's pretty small and apparently (!) it's known to be difficult
to use it as a dependency reliably.

We'll still use an external lit if we manage to find one,
but if not the shipped one should do the trick.
Not sure if this is the best route, as opposed to
forcing/standardizing on the included one.

Also include script that grabs it from a particular release script,
in case we ever want to use a newer version or something.

See:
  https://gitweb.gentoo.org/repo/gentoo.git/commit/sys-devel/llvm/files/9999?id=a1621d0c04dd6ddc4b5e4ebcef40501289292470
  http://llvm.org/PR8496