Added alld linker tool.
Added a new tool called alld, that is an allvm wrapper for the lld
linker. Also revised the strategy for chosing which linker to use
in the AOT compilation path: When lld has been built with LLVM the
out alld wrapper is getting built and is used as the default linker.
Else, the ld linker found in the PATH is used as the default.