alld: Fix linking liblldELF against latest upstream.
Upstream on longer exports liblldELF so cmake no
longer automatically adds the dependencies it needs.

We'd normally list dependencies on llvm components by
setting LLVM_LINK_COMPONENTS but that results in
the libraries appearing /before/ we have a chance
to add lldELF to the link line...

As a result we kludge our way and call llvm_config
ourselves to add dependencies on those components
(resolving them to libraries, etc.), but I don't
think this is a supported interface.