Add a simple pass to convert some inline asm into LLVM intrinsics.
After poking a bit, I found that LLVM itself has this sort of thing already
implemented--but in the TargetLowering pass, where it's only executed as part
of the lowering-to-SelectionDAG phase (which means it's not producing any actual
LLVM IR). Still, that pass does use string matching on the exact inline ASM
string for its purposes, rather than trying to assemble the instruction.

I could still go for an assemble-the-inline-asm approach, but a cursory glance
shows that actually solving the register constraints to produce parseable
assembly doesn't actually easily admit itself with LLVM's API (yet more things
buried in Target-internal land, yay). Given that previous numbers showed that
exact string matching accounts for *a lot* of the inline asm, the slightly more
liberal string matching should be more than sufficient.

No test code in this commit--I don't see any easy way to implement a lit test
that is basically running opt with the DeInlineAsm pass (which uses the new
pass manager framework because why not).