add helper that builds as override of allvm-tools from ALLVM channel.
For debugging purposes only!
This makes it easy to build in this way--
which is effectively what was sometimes done manually[1]
when using something like `nix-shell '<allvm>' -A allvm`
to create dev environment.

Not providing this top-level or in README.md since
the ALLVM channel is not (yet) public.

Curiously (investigation needed/underway) building
against the ALLVM channel avoids whatever is causing
the mass test failures currently seen in this branch.

It might have fixes or might just be getting lucky...
either way, interesting!